# swiss: optimize bucket init by setting control bytes simultaneously